Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/controller-on-disconnect.https-expected.txt


PASS controller is cleared on disconnected window
